+++ src/utils/AdBlock.ts
@@ -1,10 +1,10 @@
-import { ElectronBlocker, fullLists } from '@cliqz/adblocker-electron';
+import { ElectronBlocker, fullLists } from '@ghostery/adblocker-electron';
 import { join } from 'path';
 import { readFileSync, writeFileSync } from 'fs';
 
 export default async function (app: Electron.App, win: Electron.BrowserWindow) {
   const blocker = await ElectronBlocker.fromLists(fetch, fullLists, {
-    enableCompression: true
+    enableCompression: true,
   }, {
     path: join(app.getPath('userData'), 'engine.bin'),
     read: async (...args) => readFileSync(...args),